feat: Moving `cache` to `internal`
